x86, efi: Add whitelist for efi_variable_query_store()
Some ASUS machines require more than 50% of the EFI variable store to be in use
in order to activate garbage collection of deleted variables. Without garbage
collection it is impossible to free up space in the variable store.

We tried to avoid introducing these kinds of blacklist/whitelist but the
firmware bug spectrum is so large and varied that any other solution has the
potential to requires us to rewrite our workarounds over and over again. The
whitelist provides a natural solution to the spread of firmware bugs (which is
a consequence of the source code base fragmentation in the BIOS vendor space),
and it allows us to address bugs on a case-by-case basis.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>