efi: Capsule update support
Add the necessary interfaces to interact with the UpdateCapsule() EFI
runtime service.

Once a capsule has been passed to the firmware, the next reboot will
always be performed using the ResetSystem() EFI runtime service. This
ensures the reset value returned by QueryCapsuleCapabilities() is used
to reset the system, which is required for the capsule to be processed.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>