efi: Add support for EFI_MEMORY_RO attribute introduced by UEFIv2.5
The UEFI spec v2.5 introduces a new memory attribute EFI_MEMORY_RO,
which is now the preferred attribute to convey that the nature of
the contents of such a region allows it to be mapped read-only
(i.e., it contains .text and .rodata only). The specification of
the existing EFI_MEMORY_WP attribute has been updated to align
more closely with its common use as a cacheability attribute rather
than a permission attribute.

Add the #define and add the attribute to the memory map dumping
routine.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>