x86/efi: Map EFI_MEMORY_{XP,RO} memory region bits to EFI page tables
Now that we have EFI memory region bits that indiciate which regions
do not need execute perssion or read/write permission in the page
tables, let's use them.

A boot time analysis needs to be performed on this patch to ensure
that we haven't added a huge amount of overhead.

Additionally it would probably make a whole lotta sense to get for
EFI_PROPERTIES_TABLE and only enforce the restrictive mappings if it's
present (to allow us to ignore buggy firmware that sets bits it didn't
mean to and to preserve backwards compat).

Cc: Borislav Petkov <bp@alien8.de>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>