x86/efi: Call page table functions in kexec path
The switch to using a new page table for EFI runtime calls in commit
commit 67a9108ed431 ("x86/efi: Build our own page table structures")
failed to take into account changes required for the kexec code paths,
which are unfortunately duplicated in the EFI code.

Call the page table allocation and setup functions in
kexec_enter_virtual_mode() just like we do for
__efi_enter_virtual_mode() to avoid hitting NULL-pointer dereferences
when making EFI runtime calls.

Reported-by: Srikar Dronamraju <srikar@linux.vnet.ibm.com>
Signed-off-by: Matt Fleming <matt@codeblueprint.co.uk>