efi: Refactor efi_memmap_init() into arch-neutral code
x86 didn't provide a way to query the EFI memory map at runtime
because it was unmapped during boot in efi_free_boot_services(). This
led to the hack in efi_mem_lookup_desc() that goes and maps the memory
table by hand.

ARM and arm64 were failing to clear the EFI_MEMMAP bit when unmapping
the early EFI memory map. EFI_MEMMAP indicates whether the EFI memory
map is mapped (not the regions contained within) and can be traversed.
So it's more correct to set the bit as soon as we memremap() the
passed in EFI memmap.

While we're rename efi_unmmap_memmap() to efi_memmap_unmap() to adhere
to the regular naming scheme.

Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Leif Lindholm <leif.lindholm@linaro.org>
Cc: Peter Jones <pjones@redhat.com>
Signed-off-by: Matt Fleming <matt@codeblueprint.co.uk>