x86/efi: Always map first physical page into EFI pagetables
Commit 129766708 ("x86/efi: Only map RAM into EFI page tables if in
mixed-mode") stopped creating 1:1 mapping for all RAM in case of running
in native 64bit mode.

It turns out though that there are 64bit EFI implementations in the wild
(this particular problem has been reported on Lenovo Yoga 710-11IKB) which
still make use of first physical page for their own private use even
though they explicitly mark it EFI_CONVENTIONAL_MEMORY in the memory map.

In case there is no mapping for this particular frame in EFI pagetables,
as soon as firmware tries to make use of it, triple fault occurs and the
system reboots (in case of Yoga 710-11IKB this is very early during boot).

Fix that by always mapping the first page of physical memory into EFI
pagetables. We're on a safe side, as trim_bios_range() will reserve the
first page and isolate it away from memory allocators anyway.

Note: just reverting 129766708 is not enough on v4.9-rc1+ to fix the
regression on affected hardware, as commit ab72a27da ("x86/efi:
Consolidate region mapping logic") later made the first physical frame not
to be mapped anyway.

Reported-by: Hanka Pavlikova <hanka@ucw.cz>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Cc: stable@kernel.org # v4.8+
Cc: Waiman Long <waiman.long@hpe.com>
Cc: Borislav Petkov <bp@suse.de>
Cc: Laura Abbott <labbott@redhat.com>
Cc: Vojtech Pavlik <vojtech@ucw.cz>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Fixes: 129766708 ("x86/efi: Only map RAM into EFI page tables if in mixed-mode")
Signed-off-by: Matt Fleming <matt@codeblueprint.co.uk>